+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/messages.g.m
@@ -48,6 +48,24 @@ static id GetNullableObjectAtIndex(NSArray<id> *array, NSInteger key) {
 - (NSArray<id> *)toList;
 @end
 
+@interface FVPMediaSelectionVideoTrackData ()
++ (FVPMediaSelectionVideoTrackData *)fromList:(NSArray<id> *)list;
++ (nullable FVPMediaSelectionVideoTrackData *)nullableFromList:(NSArray<id> *)list;
+- (NSArray<id> *)toList;
+@end
+
+@interface FVPAssetVideoTrackData ()
++ (FVPAssetVideoTrackData *)fromList:(NSArray<id> *)list;
++ (nullable FVPAssetVideoTrackData *)nullableFromList:(NSArray<id> *)list;
+- (NSArray<id> *)toList;
+@end
+
+@interface FVPNativeVideoTrackData ()
++ (FVPNativeVideoTrackData *)fromList:(NSArray<id> *)list;
++ (nullable FVPNativeVideoTrackData *)nullableFromList:(NSArray<id> *)list;
+- (NSArray<id> *)toList;
+@end
+
 @implementation FVPPlatformVideoViewCreationParams
 + (instancetype)makeWithPlayerId:(NSInteger)playerId {
   FVPPlatformVideoViewCreationParams *pigeonResult =
@@ -120,6 +138,126 @@ static id GetNullableObjectAtIndex(NSArray<id> *array, NSInteger key) {
 }
 @end
 
+@implementation FVPMediaSelectionVideoTrackData
++ (instancetype)makeWithVariantIndex:(NSInteger)variantIndex
+                               label:(nullable NSString *)label
+                             bitrate:(nullable NSNumber *)bitrate
+                               width:(nullable NSNumber *)width
+                              height:(nullable NSNumber *)height
+                           frameRate:(nullable NSNumber *)frameRate
+                               codec:(nullable NSString *)codec
+                          isSelected:(BOOL)isSelected {
+  FVPMediaSelectionVideoTrackData *pigeonResult = [[FVPMediaSelectionVideoTrackData alloc] init];
+  pigeonResult.variantIndex = variantIndex;
+  pigeonResult.label = label;
+  pigeonResult.bitrate = bitrate;
+  pigeonResult.width = width;
+  pigeonResult.height = height;
+  pigeonResult.frameRate = frameRate;
+  pigeonResult.codec = codec;
+  pigeonResult.isSelected = isSelected;
+  return pigeonResult;
+}
++ (FVPMediaSelectionVideoTrackData *)fromList:(NSArray<id> *)list {
+  FVPMediaSelectionVideoTrackData *pigeonResult = [[FVPMediaSelectionVideoTrackData alloc] init];
+  pigeonResult.variantIndex = [GetNullableObjectAtIndex(list, 0) integerValue];
+  pigeonResult.label = GetNullableObjectAtIndex(list, 1);
+  pigeonResult.bitrate = GetNullableObjectAtIndex(list, 2);
+  pigeonResult.width = GetNullableObjectAtIndex(list, 3);
+  pigeonResult.height = GetNullableObjectAtIndex(list, 4);
+  pigeonResult.frameRate = GetNullableObjectAtIndex(list, 5);
+  pigeonResult.codec = GetNullableObjectAtIndex(list, 6);
+  pigeonResult.isSelected = [GetNullableObjectAtIndex(list, 7) boolValue];
+  return pigeonResult;
+}
++ (nullable FVPMediaSelectionVideoTrackData *)nullableFromList:(NSArray<id> *)list {
+  return (list) ? [FVPMediaSelectionVideoTrackData fromList:list] : nil;
+}
+- (NSArray<id> *)toList {
+  return @[
+    @(self.variantIndex),
+    self.label ?: [NSNull null],
+    self.bitrate ?: [NSNull null],
+    self.width ?: [NSNull null],
+    self.height ?: [NSNull null],
+    self.frameRate ?: [NSNull null],
+    self.codec ?: [NSNull null],
+    @(self.isSelected),
+  ];
+}
+@end
+
+@implementation FVPAssetVideoTrackData
++ (instancetype)makeWithTrackId:(NSInteger)trackId
+                          label:(nullable NSString *)label
+                          width:(nullable NSNumber *)width
+                         height:(nullable NSNumber *)height
+                      frameRate:(nullable NSNumber *)frameRate
+                          codec:(nullable NSString *)codec
+                     isSelected:(BOOL)isSelected {
+  FVPAssetVideoTrackData *pigeonResult = [[FVPAssetVideoTrackData alloc] init];
+  pigeonResult.trackId = trackId;
+  pigeonResult.label = label;
+  pigeonResult.width = width;
+  pigeonResult.height = height;
+  pigeonResult.frameRate = frameRate;
+  pigeonResult.codec = codec;
+  pigeonResult.isSelected = isSelected;
+  return pigeonResult;
+}
++ (FVPAssetVideoTrackData *)fromList:(NSArray<id> *)list {
+  FVPAssetVideoTrackData *pigeonResult = [[FVPAssetVideoTrackData alloc] init];
+  pigeonResult.trackId = [GetNullableObjectAtIndex(list, 0) integerValue];
+  pigeonResult.label = GetNullableObjectAtIndex(list, 1);
+  pigeonResult.width = GetNullableObjectAtIndex(list, 2);
+  pigeonResult.height = GetNullableObjectAtIndex(list, 3);
+  pigeonResult.frameRate = GetNullableObjectAtIndex(list, 4);
+  pigeonResult.codec = GetNullableObjectAtIndex(list, 5);
+  pigeonResult.isSelected = [GetNullableObjectAtIndex(list, 6) boolValue];
+  return pigeonResult;
+}
++ (nullable FVPAssetVideoTrackData *)nullableFromList:(NSArray<id> *)list {
+  return (list) ? [FVPAssetVideoTrackData fromList:list] : nil;
+}
+- (NSArray<id> *)toList {
+  return @[
+    @(self.trackId),
+    self.label ?: [NSNull null],
+    self.width ?: [NSNull null],
+    self.height ?: [NSNull null],
+    self.frameRate ?: [NSNull null],
+    self.codec ?: [NSNull null],
+    @(self.isSelected),
+  ];
+}
+@end
+
+@implementation FVPNativeVideoTrackData
++ (instancetype)makeWithAssetTracks:(nullable NSArray<FVPAssetVideoTrackData *> *)assetTracks
+               mediaSelectionTracks:
+                   (nullable NSArray<FVPMediaSelectionVideoTrackData *> *)mediaSelectionTracks {
+  FVPNativeVideoTrackData *pigeonResult = [[FVPNativeVideoTrackData alloc] init];
+  pigeonResult.assetTracks = assetTracks;
+  pigeonResult.mediaSelectionTracks = mediaSelectionTracks;
+  return pigeonResult;
+}
++ (FVPNativeVideoTrackData *)fromList:(NSArray<id> *)list {
+  FVPNativeVideoTrackData *pigeonResult = [[FVPNativeVideoTrackData alloc] init];
+  pigeonResult.assetTracks = GetNullableObjectAtIndex(list, 0);
+  pigeonResult.mediaSelectionTracks = GetNullableObjectAtIndex(list, 1);
+  return pigeonResult;
+}
++ (nullable FVPNativeVideoTrackData *)nullableFromList:(NSArray<id> *)list {
+  return (list) ? [FVPNativeVideoTrackData fromList:list] : nil;
+}
+- (NSArray<id> *)toList {
+  return @[
+    self.assetTracks ?: [NSNull null],
+    self.mediaSelectionTracks ?: [NSNull null],
+  ];
+}
+@end
+
 @interface FVPMessagesPigeonCodecReader : FlutterStandardReader
 @end
 @implementation FVPMessagesPigeonCodecReader
@@ -131,6 +269,12 @@ static id GetNullableObjectAtIndex(NSArray<id> *array, NSInteger key) {
       return [FVPCreationOptions fromList:[self readValue]];
     case 131:
       return [FVPTexturePlayerIds fromList:[self readValue]];
+    case 132:
+      return [FVPMediaSelectionVideoTrackData fromList:[self readValue]];
+    case 133:
+      return [FVPAssetVideoTrackData fromList:[self readValue]];
+    case 134:
+      return [FVPNativeVideoTrackData fromList:[self readValue]];
     default:
       return [super readValueOfType:type];
   }
@@ -149,6 +293,15 @@ static id GetNullableObjectAtIndex(NSArray<id> *array, NSInteger key) {
     [self writeValue:[value toList]];
   } else if ([value isKindOfClass:[FVPTexturePlayerIds class]]) {
     [self writeByte:131];
+    [self writeValue:[value toList]];
+  } else if ([value isKindOfClass:[FVPMediaSelectionVideoTrackData class]]) {
+    [self writeByte:132];
+    [self writeValue:[value toList]];
+  } else if ([value isKindOfClass:[FVPAssetVideoTrackData class]]) {
+    [self writeByte:133];
+    [self writeValue:[value toList]];
+  } else if ([value isKindOfClass:[FVPNativeVideoTrackData class]]) {
+    [self writeByte:134];
     [self writeValue:[value toList]];
   } else {
     [super writeValue:value];
@@ -496,6 +649,55 @@ void SetUpFVPVideoPlayerInstanceApiWithSuffix(id<FlutterBinaryMessenger> binaryM
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         FlutterError *error;
         [api disposeWithError:&error];
+        callback(wrapResult(nil, error));
+      }];
+    } else {
+      [channel setMessageHandler:nil];
+    }
+  }
+  /// Gets the available video tracks for the video.
+  {
+    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
+           initWithName:[NSString stringWithFormat:@"%@%@",
+                                                   @"dev.flutter.pigeon.video_player_avfoundation."
+                                                   @"VideoPlayerInstanceApi.getVideoTracks",
+                                                   messageChannelSuffix]
+        binaryMessenger:binaryMessenger
+                  codec:FVPGetMessagesCodec()];
+    if (api) {
+      NSCAssert([api respondsToSelector:@selector(getVideoTracks:)],
+                @"FVPVideoPlayerInstanceApi api (%@) doesn't respond to @selector(getVideoTracks:)",
+                api);
+      [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
+        [api getVideoTracks:^(FVPNativeVideoTrackData *_Nullable output,
+                              FlutterError *_Nullable error) {
+          callback(wrapResult(output, error));
+        }];
+      }];
+    } else {
+      [channel setMessageHandler:nil];
+    }
+  }
+  /// Selects a video track by setting preferredPeakBitRate.
+  /// Pass 0 to enable auto quality selection.
+  {
+    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
+           initWithName:[NSString stringWithFormat:@"%@%@",
+                                                   @"dev.flutter.pigeon.video_player_avfoundation."
+                                                   @"VideoPlayerInstanceApi.selectVideoTrack",
+                                                   messageChannelSuffix]
+        binaryMessenger:binaryMessenger
+                  codec:FVPGetMessagesCodec()];
+    if (api) {
+      NSCAssert([api respondsToSelector:@selector(selectVideoTrackWithBitrate:error:)],
+                @"FVPVideoPlayerInstanceApi api (%@) doesn't respond to "
+                @"@selector(selectVideoTrackWithBitrate:error:)",
+                api);
+      [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
+        NSArray<id> *args = message;
+        NSInteger arg_bitrate = [GetNullableObjectAtIndex(args, 0) integerValue];
+        FlutterError *error;
+        [api selectVideoTrackWithBitrate:arg_bitrate error:&error];
         callback(wrapResult(nil, error));
       }];
     } else {